Improve code readability

+++ .bashrc
@@ -1,19 +1,18 @@
 # ~/.bashrc: executed by bash(1) for non-login shells.
-# see /usr/share/doc/bash/examples/startup-files (in the package bash-doc)
 
-# don't put duplicate lines or lines starting with space in the history. See bash(1) for more options
+# don't put duplicate lines or lines starting with space in the history.
 HISTCONTROL=ignoreboth
 # If not running interactively, don't do anything
 case $- in
     *i*) ;;
     *) return;;
 esac
-# append to the history file, don't overwrite it
+# Append to the history file, don't overwrite it
 shopt -s histappend
-# for setting history length see HISTSIZE and HISTFILESIZE in bash(1)
+# For setting history length see HISTSIZE and HISTFILESIZE in bash(1)
 HISTSIZE=10000
 HISTFILESIZE=10000
-# check the window size after each command and, if necessary, update the values of LINES and COLUMNS.
+# Check the window size after each command and, if necessary, update the values of LINES and COLUMNS.
 shopt -s checkwinsize
 
 # If set, the pattern "**" used in a pathname expansion context will
@@ -36,9 +35,6 @@ esac
 #force_color_prompt=yes
 if [ -n "$force_color_prompt" ]; then
     if [ -x /usr/bin/tput ] && tput setaf 1 >&/dev/null; then
-        # We have color support; assume it's compliant with Ecma-48
-        # (ISO/IEC-6429). (Lack of such support is extremely rare, and such
-        # a case would tend to support setf rather than setaf.)
         color_prompt=yes
     else
         color_prompt=
@@ -59,7 +55,7 @@ case "$TERM" in
     *)
     ;;
 esac
-# colored GCC warnings and errors
+# Colored GCC warnings and errors
 export GCC_COLORS='error=01;31:warning=01;35:note=01;36:caret=01;32:locus=01:quote=01'
 # Add an "alert" alias for long running commands.  Use like so: $sleep 10; alert
 alias alert='notify-send --urgency=low -i "$([ $? = 0 ] && echo terminal || echo error)" "$(history|tail -n1|sed -e '\''s/^\s*[0-9]\+\s*//;s/[;&|]\s*alert$//'\'')"'
@@ -82,7 +78,6 @@ if ! shopt -oq posix; then
         . /etc/bash_completion
     fi
 fi
-
 
 # --------------------------------
 
@@ -154,8 +149,11 @@ mcd () {
 }
 
 # Take your .bashrc with you
-# Run it as follows: $ sshs user@server
 sshs () {
+    if [ -z "$1" ]; then 
+        echo "Usage: sshs user@server"
+        return 1
+    fi
     ssh $@ "cat > /tmp/.bashrc_temp" < ~/.bashrc_remote
     ssh -t $@ "bash --rcfile /tmp/.bashrc_temp ; rm /tmp/.bashrc_temp"
 }
